Created new services and added some validation
