+++ Service/DTOs/CreateNewUserDTO.cs
@@ -4,24 +4,14 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DAL
+namespace Service.DTOs
 {
-    public class User
+    public class CreateNewUserDTO
     {
-
-        public int UserId { get; set; }
         public string UserName { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
-        public int Age { get; set; }
         public string Email { get; set; }
         public string Password { get; set; }
-
-
-
-
-
-        public ICollection<Budget>? Budgets { get; set; } //En User kan ha många Budgets ("EN" SIDAN HÄR)
-
     }
 }
