+++ Service/DTOs/LoginDTO.cs
@@ -6,14 +6,9 @@ using System.Threading.Tasks;
 
 namespace Service.DTOs
 {
-    public class UserDTO
+    public class LoginDTO
     {
-        public int UserId { get; set; }
         public string UserName { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Email { get; set; }
-        public int Age { get; set; }
         public string Password { get; set; }
     }
 }
